adding user to db

+++ pkg/models/user.go
@@ -1,7 +1,7 @@
 package models
 
 type User struct {
-	ID       uint
+	ID       int64
 	UserName string
 	IpInfo   []IpInfo `gorm:"many2many:user_ip_infos;"`
 	Admin    bool
